+++ tests/testthat/test_MsBackendMemory.R
@@ -841,6 +841,7 @@ test_that("backendMerge,MsBackendMemory works", {
 
 test_that("selectSpectraVariables,MsBackendMemory works", {
     be <- MsBackendMemory()
+    expect_equal(peaksVariables(be), c("mz", "intensity"))
     expect_error(selectSpectraVariables(be, c("msLevel", "other")),
                  "not available")
 
@@ -848,11 +849,20 @@ test_that("selectSpectraVariables,MsBackendMemory works", {
     expect_error(selectSpectraVariables(be, c("msLevel")), "dataStorage")
     res <- selectSpectraVariables(be, c("msLevel", "dataStorage"))
     expect_equal(colnames(res@spectraData), c("msLevel", "dataStorage"))
-    expect_equal(peaksVariables(res), NULL)
+    expect_equal(peaksVariables(res), c("mz", "intensity"))
+    expect_s4_class(intensity(res), "NumericList")
+    expect_s4_class(mz(res), "NumericList")
+    expect_true(length(res@peaksData) == 3)
+    expect_true(all(vapply(res@peaksData, is.matrix, logical(1))))
+    expect_true(all(vapply(res@peaksData, nrow, integer(1)) == 0))
 
     res <- selectSpectraVariables(be, c("mz", "dataStorage"))
     expect_equal(colnames(res@spectraData), c("dataStorage"))
     expect_equal(peaksVariables(res), "mz")
+    expect_equal(mz(res), mz(be))
+    expect_s4_class(intensity(res), "NumericList")
+    expect_equal(length(intensity(res)), length(mz(res)))
+    expect_true(all(lengths(intensity(res)) == 0))
 
     tmp <- test_df
     tmp$peak_anno <- list(c("a", "", "b"), c("a", "b"), c("a", "b", "c", "d"))